requirements.txt
streamlit
pandas 
numpy 
requests
Pillow
opencv-python
tqdm
jupyterlab